+++ gdporch/src/main/java/it/csipiemonte/gdp/gdporch/model/entity/GdpDataUscita.java
@@ -13,10 +13,10 @@ public class GdpDataUscita extends PanacheEntityBase {
 
     @Id
     @Column(name = "ID_GDP_DATA_USCITA")
-    public Long id;
+    public Integer id;
 
     @Column(name = "FK_GDP_PERIODICITA", nullable = false)
-    public Long fkGdpPeriodicita;
+    public Integer fkGdpPeriodicita;
 
     @Column(name = "DT_INIZIO", nullable = false)
     public LocalDate dataInizio;
