+++ gdporch/src/main/java/it/csipiemonte/gdp/gdporch/model/entity/GdpEdizione.java
@@ -13,10 +13,10 @@ public class GdpEdizione extends PanacheEntityBase {
 
     @Id
     @Column(name = "ID_GDP_EDIZIONE")
-    public Long id;
+    public Integer id;
 
     @Column(name = "FK_GDP_TESTATA", nullable = false)
-    public Long fkGdpTestata;
+    public Integer fkGdpTestata;
 
     @Column(name = "DATA_EDIZIONE", nullable = false)
     public LocalDate dataEdizione;
